% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{data_cleaning}
\alias{data_cleaning}
\title{Template for data cleaning}
\usage{
data_cleaning(report_month, report_year)
}
\arguments{
\item{report_month}{Month of report}

\item{report_year}{Year of report}
}
\description{
Opens a template of steps for cleaning the month's delays data.
}
