% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delays_report.R
\name{delays_report}
\alias{delays_report}
\title{Template for creating delays report}
\usage{
delays_report(report_month, report_year, render = FALSE)
}
\arguments{
\item{report_month}{Month of report}

\item{report_year}{Year of report}

\item{render}{Should the report be rendered automatically? Defaults to `FALSE`}
}
\description{
Opens a template of the delays report for the month, optionally rendering it.
}
