% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Set up delays report}
\usage{
setup(report_month, report_year)
}
\arguments{
\item{report_month}{Month of report}

\item{report_year}{Year of report}
}
\description{
Set up delays report, with instructions on how to get the data and helper functions to clean the data and write the report.
}
